backend/requirements.txt
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
statsmodels==0.14.2
joblib==1.4.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
